Refactor: Move auth.rs and storage.rs to main src directory to prepare for client crate migration
- Move auth.rs and storage.rs from src/cli/ to src/ for better module organization
- Update all imports to use new top-level module paths
- Add auth and storage modules to main.rs
- Remove old module declarations from cli.rs
- Maintain all functionality while eliminating circular dependencies
- Prepare foundation for client crate extraction
