+++ crates/meroctl/src/config.rs
@@ -7,10 +7,10 @@ use serde::{Deserialize, Serialize};
 use tokio::fs;
 use url::Url;
 
-use crate::cli::storage::JwtToken;
 use crate::common::{fetch_multiaddr, load_config, multiaddr_to_url};
 use crate::connection::ConnectionInfo;
 use crate::output::Output;
+use crate::storage::JwtToken;
 
 #[derive(Debug, Default, Serialize, Deserialize)]
 pub struct Config {
